Codificacion en capa de vista, ralizacion de la tabla de unidades de aprendizaje
