+++ desarrollo/vista/src/main/java/mx/desarrollo/ui/UnidadesAprendizajeUI.java
@@ -13,27 +13,45 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import javax.inject.Named;
 import java.util.List;
+import javax.faces.bean.ManagedBean;
 /**
  *
  * @author PC
  */
-@Named(value = "unidadesAprendizajeUI")
+@ManagedBean (name= "UAUI")
 @SessionScoped
 public class UnidadesAprendizajeUI implements Serializable {
     private List<Learningunit> learningunits;
-//    private Learningunit learningunit''
-    private learningUnitHelper helper;
+    private Learningunit learningunit;
+    private final learningUnitHelper helper;
 
     public UnidadesAprendizajeUI() {
-        learningunits = new ArrayList();
+        helper = new learningUnitHelper();
+        this.learningunits = helper.getLearningUnits();
     }
-    @PostConstruct
-    public void init(){
-        learningunits = helper.getLearningUnits();
-    }
-    
+
     public List<Learningunit> getUnidadAprendizaje() {
         return learningunits;
+    }
+
+    public List<Learningunit> getLearningunits() {
+        return learningunits;
+    }
+
+    public void setLearningunits(List<Learningunit> learningunits) {
+        this.learningunits = learningunits;
+    }
+
+    public Learningunit getLearningunit() {
+        return learningunit;
+    }
+
+    public void setLearningunit(Learningunit learningunit) {
+        this.learningunit = learningunit;
+    }
+    
+    public void saveUnidadAprendizaje(){
+        
     }
     
 }
