+++ desarrollo/vista/src/main/webapp/altaProfesor.xhtml
@@ -13,153 +13,9 @@ and open the template in the editor.
       xmlns:f="http://xmlns.jcp.org/jsf/core">
     <h:head>
         <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'/>
-
+        <link href="./resources/style.css" rel="stylesheet"/>
         <title>Alta de profesores</title>
-        
-        <style>
-            * {
-                margin: 0;
-                padding: 0;
-                box-sizing: border-box;
-                font-family: sans-serif;
-            }
-            
-            body {
-                background-color: #EBEBEB;
-                width: 100%;
-                height: 100vh;
-            }
-            
-            .contenedorPrincipal {
-                display: flex;
-                flex-direction: row;
-                width: 100%;
-                height: 100%;
-            }
-            
-            .leftSide-contenedor {
-                width: 2.3cm;
-                height: 100%;
-                background: none;
-                position: absolute;
-                overflow: hidden;
-                transition: width 0.1s ease-in-out;
-            }
-            
-            .leftSide {
-                display: flex;
-                flex-direction: column;
-                align-items: left;
-                width: 10cm;
-                background: #007141;
-                height: 100vh;
-                position: absolute;
-                z-index: 1;
-                padding: 20px;
-            }
-            
-            .contenidoLeft {
-                display: flex;
-                flex-direction: row;
-                align-items: center;
-                width: 100%;
-                margin-bottom: 1cm;
-            }
-            
-            .leftSide-contenedor:hover {
-                width: 10cm;
-            }
-            
-            .Forma {
-                width: 100%;
-                height: 90%;
-                padding: 1cm;
-                background-color: white;
-                margin: 1cm 1cm 1cm 4cm;
-                border-radius: 20px;
-                box-shadow: 15px 15px rgba(0, 0, 0, 0.2);
-                display: flex;
-                flex-direction: column;
-                justify-content: space-between;
-            }
-            
-            h2 {
-                font-weight: normal;
-                font-size: 1.6cm;
-            }
-            
-            h3 {
-                font-size: 0.6cm;
-                margin-left: 20px;
-                color: white;
-            }
-            
-            a {
-                text-decoration: none;
-            }
-            
-            .Forma .campo {
-                display: flex;
-                flex-direction: column;
-                margin-bottom: 1cm
-            }
-            
-            .Forma .labelText {
-                margin-bottom: 15px;
-                font-size: 24px;
-            }
-            
-            .Forma .textInput {
-                width: 8cm;
-                height: 1cm;
-                margin-bottom: 45px;
-                border: 2px solid #EBEBEB;
-                border-radius: 10px;
-                padding: 15px;
-                box-shadow: none;
-            }
-            
-            .textInput:hover{
-                border: 2px solid #007141;
-            }
-            
-            .botones {
-                display: flex;
-                flex-direction: row;
-                justify-content: space-between;
-                width: 8cm;
-            }
-            .checkMenu {
-                padding: 10px;
-                background: white;
-                border: 2px solid #EBEBEB;
-                border-radius: 10px;
-            }
-            
-            .Forma .boton {
-                border: none;
-                width: 3.5cm;
-                height: 1.5cm;
-                font-weight: normal;
-                border-radius: 10px;
-                text-decoration: none;
-                box-shadow: none;
-            }
-            
-            .subForma {
-                height: 70%;
-            }
-            
-            .col-1 {
-                background: #007141;
-                color: white
-            }
-            
-            .col-2 {
-                background: red;
-                color: white
-            }
-        </style>
+       
     </h:head>
     
     <h:body>
@@ -225,7 +81,7 @@ and open the template in the editor.
                 
             <div class="campo">
                 <p:selectCheckboxMenu label="Unidades de Aprendizaje" updateLabel="true" id="seleccion" filter="true" filterMatchMode="startsWith"
-                                      style="width: 8cm" panelStyle="width: 8cm" scrollHeight="250" class="checkMenu" value="#{Profesor.unidadesDeAprendizaje}">
+                                      style="width: 8cm" panelStyle="width: 8cm" scrollHeight="250" class="checkMenu" value="#{UAUI.learningunits}">
                 </p:selectCheckboxMenu>
             </div>
