+++ desarrollo/vista/src/main/webapp/consultaUnidadesDeAprendizaje.xhtml
@@ -61,28 +61,28 @@ and open the template in the editor.
             <h:form class="Forma">
                 <h2>Consulta de Unidades De Aprendizaje</h2>
                 
-                <p:dataTable class="tabla" scrollable='true' value="#{unidadesAprendizajeUI.unidadAprendizaje}" var="value">
+                <p:dataTable class="tabla" scrollable='true' value="#{UAUI.learningunits}" var="learningunit">
                     <p:column width="20%">
                         <f:facet name="header">ID</f:facet>
-                        
+                        #{learningunit.id}
                     </p:column>
                     <p:column width="50%">
                         <f:facet name="header">Unidad de Aprendizaje</f:facet>
-                        #{value.name}
+                        #{learningunit.name}
                     </p:column>
                     <p:column width="10%">
                         <f:facet name="header">Horas de Clase</f:facet>
-                        #{value.classHours}
+                        #{learningunit.classHours}
 
                     </p:column>
                     <p:column width="10%">
                         <f:facet name="header">Horas de Taller</f:facet>
-                        #{value.workshopHours}
+                        #{learningunit.workshopHours}
 
                     </p:column>
                     <p:column width="10%">
                         <f:facet name="header">Horas de Laboratorio</f:facet>
-                        #{value.laboratoryHours}
+                        #{learningunit.laboratoryHours}
                     </p:column>
                 </p:dataTable>
